requirements.txt
# Google Calendar API
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.110.0

# macOS EventKit access
pyobjc-core==10.1
pyobjc-framework-EventKit==10.1

# Configuration and utilities
PyYAML==6.0.1

# Database
# (using SQLite - built-in to Python)

# CLI and logging
# (using argparse and logging - built-in to Python)
